requirements.txt
networkx
torch
groq
torchvision
torch-summary
einops
scikit-image
opencv-python
pyclipper
shapely
requests
cryptography
freetype-py
aiohttp
tqdm
deepl
ImageHash
kornia
backports.cached-property
huggingface_hub
transformers
py3langid==0.2.2
sentencepiece
editdistance
numpy==1.26.4
tensorboardX
websockets
protobuf
ctranslate2<=3.24.0 # last version compatible with CUDA 11
colorama
openai==1.35.9
httpx==0.27.2 # stop before blocking change in 0.28.0
open_clip_torch
safetensors
pandas
onnxruntime
timm
omegaconf
python-dotenv
nest-asyncio
marshmallow
cython
aioshutil
aiofiles
arabic-reshaper
pyhyphen
langcodes
manga-ocr
langdetect
pydensecrf@https://github.com/lucasb-eyer/pydensecrf/archive/refs/heads/master.zip
accelerate
bitsandbytes
uvicorn
fastapi
pydantic
python-multipart
paddleocr
paddlepaddle==2.5.2
# paddlepaddle-gpu; sys_platform != 'darwin'
asyncpg
SQLAlchemy
databases
alembic
pydantic_settings
psycopg2
aioredis
python-multipart
boto3
fastapi_users
fastapi-users[sqlalchemy]
fastapi_users_db_sqlalchemy
httpx
ulid-py
